+++ earthfile2llb/earthfile2llb.go
@@ -4,6 +4,11 @@ import (
 	"context"
 	"fmt"
 
+	"github.com/moby/buildkit/client/llb"
+	gwclient "github.com/moby/buildkit/frontend/gateway/client"
+	specs "github.com/opencontainers/image-spec/specs-go/v1"
+	"github.com/pkg/errors"
+
 	"github.com/earthly/earthly/ast"
 	"github.com/earthly/earthly/buildcontext"
 	"github.com/earthly/earthly/buildcontext/provider"
@@ -12,10 +17,6 @@ import (
 	"github.com/earthly/earthly/llbutil"
 	"github.com/earthly/earthly/states"
 	"github.com/earthly/earthly/variables"
-	"github.com/moby/buildkit/client/llb"
-	gwclient "github.com/moby/buildkit/frontend/gateway/client"
-	specs "github.com/opencontainers/image-spec/specs-go/v1"
-	"github.com/pkg/errors"
 )
 
 // ConvertOpt holds conversion parameters.
@@ -113,9 +114,6 @@ func Earthfile2LLB(ctx context.Context, target domain.Target, opt ConvertOpt) (m
 	}
 	converter, err := NewConverter(ctx, bc.Ref.(domain.Target), bc, opt)
 	if err != nil {
-		return nil, err
-	}
-	if err := validateAst(bc.Earthfile); err != nil {
 		return nil, err
 	}
 	interpreter := newInterpreter(converter, opt.Resolver, opt.GwClient, target)
